Added AudioManager that's tied with Audio Mixer. Using audio groups.
I added an AudioManager. In this AudioManager, we can add all the sounds to it. From this AudioManager, we can call Play(name), or Play(name, volume, pitch) for any sound. For some sounds, such as those in the "pressure" group, their volume and pitch is slightly randomized to add variance. For now, I've also added AudioTester which responds to spacebar, numpads 0-9, and up and down arrows just to make sure all the audio works

To-Do: check if an audio source is already playing so I can toggle or disable toggle.